# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        description: Error message
        type: string
    type: object
...
  models.Product:
    properties:
      brand:
        type: string
      category:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      size:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.PromoCode:
    properties:
      active:
        type: boolean
      code:
        type: string
      description:
        type: string
      discount_percentage:
        type: number
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      times_used:
        type: integer
      usage_limit:
        type: integer
    type: object
  models.Promotion:
    properties:
      created_at:
        type: string
      description:
        type: string
      discount:
        type: number
      end_date:
        type: string
      id:
        type: integer
      image_url:
        type: string
      notify_users:
        type: boolean
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Trainer:
    properties:
      about:
        type: string
      age:
        type: integer
      country:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      height:
        type: number
      id:
        type: integer
      image_url:
        type: string
      trainer_type:
        type: string
      updated_at:
        type: string
      weight:
        type: number
    type: object
...
  models.TrainingSession:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      num_places:
        type: integer
      start_time:
        type: string
      trainer_id:
        type: integer
      training_date:
        type: string
      training_type:
        type: string
      updated_at:
        type: string
    type: object
...
host: 25.28.121.106:8080
info:
  contact: {}
  description: API для Basketball Academy
  title: Basketball Academy API
  version: "1.0"
paths:
...
  /api/Products:
    get:
      consumes:
      - application/json
      description: Get a list of all Products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all Products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new Product with the given information
      parameters:
      - description: Product data
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new Product
      tags:
      - Products
  /api/Products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete Product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a Product by their ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update Product data by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Product data
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update Product
      tags:
      - Products
  /api/PromoCodes:
    get:
      consumes:
      - application/json
      description: Get a list of all PromoCodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PromoCode'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all PromoCodes
      tags:
      - PromoCodes
    post:
      consumes:
      - application/json
      description: Create a new PromoCode with the given information
      parameters:
      - description: PromoCode data
        in: body
        name: PromoCode
        required: true
        schema:
          $ref: '#/definitions/models.PromoCode'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PromoCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new PromoCode
      tags:
      - PromoCodes
  /api/Promotions:
    get:
      consumes:
      - application/json
      description: Get a list of all Promotions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Promotion'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all Promotions
      tags:
      - Promotions
    post:
      consumes:
      - application/json
      description: Create a new Promotion with the given information
      parameters:
      - description: Promotion data
        in: body
        name: Promotion
        required: true
        schema:
          $ref: '#/definitions/models.Promotion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Promotion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new Promotion
      tags:
      - Promotions
  /api/Promotions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Promotion by ID
      parameters:
      - description: Promotion ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete Promotion
      tags:
      - Promotions
    put:
      consumes:
      - application/json
      description: Update Promotion data by ID
      parameters:
      - description: Promotion ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Promotion data
        in: body
        name: Promotion
        required: true
        schema:
          $ref: '#/definitions/models.Promotion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Promotion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update Promotion
      tags:
      - Promotions
...
  /api/trainers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Trainer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить всех тренеров
      tags:
      - Trainers
  /api/trainers/{id}:
    get:
      parameters:
      - description: ID тренера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trainer'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить тренера по ID
      tags:
      - Trainers
...
  /api/training-sessions/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID тренировки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные тренировки
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.TrainingSession'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TrainingSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить данные тренировки (только для тренера)
      tags:
      - Trainers
...
swagger: "2.0"